+++ src/pyhgf/updates/categorical.py
@@ -4,11 +4,9 @@ from functools import partial
 from typing import Dict
 
 import jax.numpy as jnp
-from jax import Array, jit
-from jax.scipy.special import digamma, gamma
-from jax.typing import ArrayLike
+from jax import jit
 
-from pyhgf.math import binary_surprise
+from pyhgf.math import binary_surprise, dirichlet_kullback_leibler
 from pyhgf.typing import Edges
 
 
@@ -111,46 +109,3 @@ def categorical_input_update(
     attributes[node_idx]["time_step"] = time_step
 
     return attributes
-
-
-def dirichlet_kullback_leibler(alpha_1: ArrayLike, alpha_2: ArrayLike) -> Array:
-    r"""Compute the Kullback-Leibler divergence between two Dirichlet distributions.
-
-    The Kullback-Leibler divergence from the distribution :math:`Q` to the distribution
-    :math:`P`, two Dirichlet distributions parametrized by :math:`\\alpha_2` and
-    :math:`\\alpha_1` (respectively) is given by the following equation:
-
-    .. math::
-       KL[P||Q] = \\
-       \\ln{\\frac{\\Gamma(\\sum_{i=1}^k\\alpha_{1i})} \\
-       {\\Gamma(\\sum_{i=1}^k\\alpha_{2i})}} + \\
-       \\sum_{i=1}^k \\ln{\\frac{\\Gamma(\\alpha_{2i})}{\\Gamma(\\alpha_{1i})}} + \\
-       \\sum_{i=1}^k(\\alpha_{1i} -\\
-       \\alpha_{2i})\\left[\\psi(\alpha_{1i})-\\psi(\\sum_{i=1}^k\alpha_{1i})\\right]
-
-    Parameters
-    ----------
-    alpha_1 :
-        The concentration parameters for the distribution :math:`P`.
-    alpha_2 :
-        The concentration parameters for the distribution :math:`Q`.
-
-    Returns
-    -------
-    kl :
-        The Kullback-Leibler divergence of distribution :math:`P` from distribution
-        :math:`Q`.
-
-    References
-    ----------
-    .. [1] https://statproofbook.github.io/P/dir-kl.html
-    .. [2] Penny, William D. (2001): "KL-Divergences of Normal, Gamma, Dirichlet and
-       Wishart densities" ; in: University College, London , p. 2, eqs. 8-9 ;
-       URL: https://www.fil.ion.ucl.ac.uk/~wpenny/publications/densities.ps .
-
-    """
-    return (
-        jnp.log(gamma(alpha_1.sum()) / gamma(alpha_2.sum()))
-        + jnp.sum(jnp.log(gamma(alpha_2) / gamma(alpha_1)))
-        + jnp.sum((alpha_1 - alpha_2) * (digamma(alpha_1) - digamma(alpha_1.sum())))
-    )
